add more integration tests for query + fix detected bugs
