+++ src/modules/ogcapi-records/ogcapi-records-implementation/src/main/java/org/geonetwork/ogcapi/records/OgcapiCollectionsApiController.java
@@ -107,7 +107,7 @@ public class OgcapiCollectionsApiController implements CollectionsApi {
             Integer offset,
             List<String> q,
             List<String> type,
-            List<String> externalId,
+            List<String> externalIds,
             List<String> ids,
             List<String> sortby,
             String filter,
@@ -122,7 +122,7 @@ public class OgcapiCollectionsApiController implements CollectionsApi {
                 type,
                 q,
                 ids,
-                externalId,
+                externalIds,
                 sortby,
                 filter,
                 filterLang,
