+++ src/modules/ogcapi-records/ogcapi-records-openapi-autogen/src/ogc-openapi-schema/openapi/parameters/externalId.yaml
@@ -1,5 +1,5 @@
 ---
-name: externalId
+name: externalIds
 in: query
 description: |-
   The optional externalId parameter supports searching by an identifier
